% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docco.R
\docType{data}
\name{airlines}
\alias{airlines}
\title{Flights delayed and on time for 2 airlines and 5 airports}
\format{A data frame with 5 rows and 4 columns:
\describe{
  \item{airport}{Name of airport}
  \item{aa.ontime}{Frequency of Alaska Airlines flights arriving at each airport
  that were on time}
  \item{aa.delayed}{Frequency of Alaska Airlines flights arriving at each airport
  that were delayed}
  \item{aw.ontime}{Frequency of America West flights arriving at each airport that
  were on time}
  \item{aw.delayed}{Frequency of America West flights arriving at each airport that
  were delayed}
}}
\source{
Moore, Basic Practice of Statistics.
}
\usage{
data(airlines)
}
\description{
A dataset containing numbers of flights delayed and on time at 5 different US West
Coast airports for 2 different airlines (Alaska Airlines, "aa" and America West "aw").
}
\keyword{datasets}

