% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docco.R
\docType{data}
\name{ancova}
\alias{ancova}
\title{Sample data for analysis of covariance}
\format{A data frame with 20 rows and 3 columns:
\describe{
  \item{drug}{Name of drug, "a" or "b"}
  \item{before}{Subject's score before being given the drug}
  \item{after}{Subject's score after receiving the drug}
}}
\source{
Dunno, Zar?
}
\usage{
data(ancova)
}
\description{
Before and after scores for subjects given two different drugs.
}
\keyword{datasets}

