% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docco.R
\docType{data}
\name{autonoise}
\alias{autonoise}
\title{Automobile filter noise data}
\format{A data frame with 36 rows and 4 variables:
\describe{
  \item{noise}{Noise level (decibels)}
  \item{size}{Size of car being tested: Large, Medium or Small}
  \item{type}{Type of filter, Standard or Octel}
  \item{side}{Side of car, Left or Right}
}}
\source{
Mendenhall and Sincich
}
\usage{
data(autonoise)
}
\description{
Data gathered to assess whether the silencing properties of the Octel filter were
"at least equal to standard filters".
}
\keyword{datasets}

